Implement Italian Player Words
Add to library `lib_supplement.i`, a temporary library module that deals with
creating Italian synonyms for Predefined Player Words. This will be removed
once Italian language gets natively supported in Alan; in the meanwhile it
will serve as a test ground for deciding which Player Words should be hardcoded
in the final Italian native support.