Lib v0.1.0: Preposizioni Articolate
This commit introduces major changes in the library, therefore all library files are bumped up to v0.1.0.

A new string attribute `articolo` is now added to EVERY THING, used by authors to store the definite article of the instance's noun. The library will use this value to automatically set gender, number, DEFINITE/INDEFINITE ARTICLE, and the preposition forms (preposizioni articolate) that match the noun.

New attributes are added to EVERY THING to store the prepositions:

    HAS prep_DI "del".
    HAS prep_A  "al".
    HAS prep_DA "dal".
    HAS prep_IN "nel".
    HAS prep_SU "sul".