Who/Where Verbs
This commit translates the group of `who/where_*` verbs that deal with questions like "who am I"
typed by the player. During translation a bug in Alan was unveiled which prevents usage of "è" in
SYNTAX and SYNONYM statements (while it works fine with instances). All other Italian accented
letters work fine, so in three of the translated verbs the "é" had to be used instead of "è" (which is
grammatically incorrect, but will have to do for the meantime).