New THING Attribute `vocale`
Add the new `vocale` string attribute to every `THING` instance in order for
the library to better handle conjucation of adjectives in library messages.

This new features reduces considerably the size of VERB definitions, and makes
creation of custom verbs and messages much easier on the authors' side.

All library files are bumped up to version `0.4.0`.