Turning On/Off Devices and Lightsources
This commit translates and adapts to Italian the group of verbs and attributes
that deal with switching on/of devices and lightsources. It's a big change that
envolves structural changes to the library verbs and attributs because of the
differences in the two languages regarding the use of these verbs.

Some original English verbs have been removed and the Italian verbs `accendi`
and `spegni` (turn on/off) have been implemented to handle both devices and
lightsources alike. The `natural` attribute for lightsources was dropped as it
didn't serve any purpose after adapation of the verbs.