Verb Search + Fix 'Ascolta'
Verb restriction `search` to `ispezionare`.

Verb `search` to `ispeziona`.

Fix `ascolta`: In `DOES` change `IF ogg AT hero` into `IF hero AT ogg`
in order for the verb to work when hero is in a nested location (eg. in
a hole) and to work with room/site objects (which are not AT hero's
location because they're in a wrapping location, but the hero is AT their
location).